requirements.txt
arrow==0.13.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-SocketIO==3.3.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.1
logzero==1.5.0
MarkupSafe==1.1.1
peewee==3.9.5
python-dateutil==2.8.0
python-engineio==3.5.1
python-socketio==3.1.2
qiniu==7.2.2
requests==2.21.0
six==1.12.0
urllib3==1.24.3
websocket-client==0.56.0
Werkzeug==0.15.2
wtf-peewee==3.0.0
WTForms==2.2.1
